HIA-1277: Add feature flag to send decimals (#1563)
* HIA-1277: Add feature flag to send decimals

* HIA-1277: Add feature flag to send decimals

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/assessRisksAndNeeds/ArnSexualPredictorScore.kt
@@ -1,12 +1,13 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.assessRisksAndNeeds
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.SexualPredictor
+import java.math.BigDecimal
 
 data class ArnSexualPredictorScore(
   val ospIndecentScoreLevel: String? = null,
   val ospContactScoreLevel: String? = null,
-  val ospIndirectImagePercentageScore: Int? = null,
-  val ospDirectContactPercentageScore: Int? = null,
+  val ospIndirectImagePercentageScore: BigDecimal? = null,
+  val ospDirectContactPercentageScore: BigDecimal? = null,
 ) {
   fun toSexualPredictor(useV2NumericalValue: Boolean = true): SexualPredictor =
     SexualPredictor(
